Clacks Tower structures and cards
Defined simple examples of three clacks towers using a map for each
tower.  Each map has two key/value pairs, each pair representing a row
of light patterns to display.

Defined a very simple card for each of the towers to display the value of the light
patterns as 1's and 0's.